code/requirements.txt
pydantic==1.10.8
pytest==7.4.0
pytest-mock==3.11.1
requests==2.31.0
superagi-tools==1.0.6
tiktoken==0.4.0
typing_extensions==4.7.1
urllib3==1.26.16
filelock==3.12.0
requests-file==1.5.1
watchfiles==0.19.0
